Add random color to each boxes/
At first the btn node has been selected. Later a boolean variable of 'isRandomColor' has been taken. On click of the button the value will be changed to true or false. And random colors will be created as inline css. This random color is created from in "line 23". And is applied in "line 26".